+++ Sources/LightstreamerClient/utils/Constants.swift
@@ -1,16 +1,16 @@
 import Foundation
 
 #if os(macOS)
-var LS_CID = "scFuxkwp1ltvcB4BJ4JikvD8h"
+var LS_CID = "scFuxkwp1ltvcB4BJ4JikvD9i"
 #elseif os(tvOS)
-var LS_CID = "zxRyen2m uz.k55AK1xlzchl"
+var LS_CID = "zxRyen2m uz.k55AK1xlzcim"
 #elseif os(watchOS)
-var LS_CID = ".cWimz9dysogQz2GJ4L73dHzfDm"
+var LS_CID = ".cWimz9dysogQz2GJ4L73dHzfEn"
 #else
-var LS_CID = "oqVfhw.i6 37e64BHf  g4h"
+var LS_CID = "oqVfhw.i6 37e64BHf  g5i"
 #endif
 let LS_LIB_NAME = "swift_client"
-let LS_LIB_VERSION = "5.0.0 beta2"
+let LS_LIB_VERSION = "5.0.0 beta3"
 
 var TLCP_VERSION = "TLCP-2.3.0"
 let FULL_TLCP_VERSION = TLCP_VERSION + ".lightstreamer.com"